Rename WatchStats.App project to WatchStats.Cli
