+++ WatchStats.Cli/AssemblyInfo.cs
