+++ WatchStats.Cli/WatchStats.Cli.csproj
@@ -5,7 +5,6 @@
         <TargetFramework>net10.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>WatchStats.App</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
